+++ lib/src/scope/g_lite_scope/lite_scope_base.dart
@@ -25,6 +25,13 @@ abstract base class LiteScope<W extends LiteScope<W, S>,
   /// A callback invoked when the [initCancellationTimeout] expires.
   final void Function()? onInitCancellationTimeout;
 
+  /// The timeout duration for the asynchronous teardown of this scope, before
+  /// it triggers [onDisposeAsyncTimeout].
+  final Duration? disposeAsyncTimeout;
+
+  /// A callback invoked when the [disposeAsyncTimeout] expires.
+  final void Function()? onDisposeAsyncTimeout;
+
   /// The timeout duration for waiting to dispose child scopes.
   final Duration? waitForChildrenTimeout;
 
@@ -43,6 +50,8 @@ abstract base class LiteScope<W extends LiteScope<W, S>,
     this.onScopeKeyTimeout,
     this.initCancellationTimeout,
     this.onInitCancellationTimeout,
+    this.disposeAsyncTimeout,
+    this.onDisposeAsyncTimeout,
     this.waitForChildrenTimeout,
     this.onWaitForChildrenTimeout,
     this.pauseAfterInitialization,
@@ -189,6 +198,12 @@ final class _LiteScopeElement<W extends LiteScope<W, S>,
 
   @override
   void onInitCancellationTimeout() => widget.onInitCancellationTimeout?.call();
+
+  @override
+  Duration? get disposeAsyncTimeout => widget.disposeAsyncTimeout;
+
+  @override
+  void onDisposeAsyncTimeout() => widget.onDisposeAsyncTimeout?.call();
 
   @override
   Duration? get waitForChildrenTimeout => widget.waitForChildrenTimeout;